requirements.txt
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
mutagen==1.45.1
Pillow==9.2.0
pycryptodomex==3.15.0
pyserial==3.5
requests==2.28.1
urllib3==1.26.11
websockets==10.3
youtube-search==2.1.1
yt-dlp==2022.8.8
